Extract constellations names
Constellation names are now extracted from java to xml string
resources, which makes them ready to be translated!

Constellation abbreviations are now acting as keys to load the strings,
so they should always be in the same format. That is: IAU designation
all in capital letters.

Abbreviations are still hard coded in places.